% Generated by roxygen2 (4.0.1): do not edit by hand
\name{char2bool}
\alias{char2bool}
\title{char2bool}
\usage{
char2bool(x, var_names, CR_as_TRUE = FALSE, NA_as_FALSE = FALSE)
}
\arguments{
\item{x}{data frame}

\item{var_names}{variable to be converted}

\item{NA_as_FALSE}{convert \code{NA} into \code{FALSE}}

\item{CR_as_TRUE}{logical, if \code{TRUE}, CR (can't remember) is replaced by \code{TRUE} and \code{FALSE} otherwise. This choice is required for the question of whether current illness is the same bout as the one reported the previous time.}
}
\description{
char2bool
}

