% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean_intake_survey}
\alias{clean_intake_survey}
\title{Clean intake survey}
\usage{
clean_intake_survey(flunet, age_max_children = 10)
}
\arguments{
\item{age_max_children}{numeric. Upper age limit for children. Below this age, profile with "adult" behaviour (smoking, employment) are flagged as suspicious.}

\item{flunet}{a \code{\link{flunet}} object}
}
\description{
This function performs several checks, resolves them when possible and otherwise flags informative warnings on the corresponding profiles.
}

