% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean_profile}
\alias{clean_profile}
\title{Resolve multiple profile entries for Flusurvey participants. Rename variables and cast values in R format.}
\usage{
clean_profile(df_profile, age_group_cut = c(0, 18, 45, 65, Inf),
  age_group_name = c("0-17", "18-44", "45-64", "65+"))
}
\arguments{
\item{df_profile}{a \link{data.frame} containing FluSurvey participants profiles.}

\item{age_group_cut}{a numeric vector containing the start and end points of the age groups.}

\item{age_group_name}{a character vector containing the labels for the age groups.}
}
\value{
a \link{data.frame}.
}
\description{
Resolve multiple profile entries for Flusurvey participants. Rename variables and cast values in R format.
}

