% Generated by roxygen2 (4.0.1): do not edit by hand
\name{count_same_bout}
\alias{count_same_bout}
\title{Group reports by same bout}
\usage{
count_same_bout(df_weekly, CR_as_TRUE = FALSE, give_position = FALSE,
  give_length = FALSE, subset = NULL)
}
\arguments{
\item{df_weekly}{\code{data.frame} containing weekly survey data for a single participant.}

\item{give_position}{logical, if TRUE the position within each bout is returned in the column "position_bout"}

\item{give_length}{logical, if TRUE the length of each bout is returned in the column "length_bout"}

\item{CR_as_TRUE}{logical, if \code{TRUE}, CR (can't remember) is replaced by \code{TRUE} and \code{FALSE} otherwise. This choice is required for the question of whether current illness is the same bout as the one reported the previous time.}

\item{subset}{character, logical expression indicating reports to keep: missing values are taken as false. If present, only episode of illness with at least one report that verify \code{subset} will be processed. This is mainly to save time. E.g. to clean only episodes with at least one ARI report, one can uses \code{subset="ARI_ecdc"}.}
}
\description{
Group reports by same bout
}
\note{
this function account for multiple report on the same date by grouping them in the same bout.
}

