% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_context_flusurvey_201213}
\alias{create_context_flusurvey_201213}
\title{context of the 2012/2013 Flusurvey season}
\usage{
create_context_flusurvey_201213()
}
\value{
list of two items:
\enumerate{
	\item context_R: list of context variables in R format
	\item context_JSON: list of context variables in JSON format
}
}
\description{
This function create the context of the 2012/2013 Flusurvey season
}

