\name{create_dictionnary_flusurvey_201213}
\alias{create_dictionnary_flusurvey_201213}
\title{Dictionary of the 2012/2013 Flusurvey season}
\usage{
create_dictionnary_flusurvey_201213()
}
\value{
list of two items: \enumerate{ \item dico_R: list of dico
variables in R format \item dico_JSON: list of dico
variables in JSON format }
}
\description{
This function create the dictionary of the 2012/2013
Flusurvey season
}

