% Generated by roxygen2 (4.0.1): do not edit by hand
\name{empirical distribution}
\alias{empirical distribution}
\alias{empirical_distribution_symptom_duration}
\alias{empirical_distribution_time_to_report}
\title{Compute empirical distributions}
\usage{
empirical_distribution_symptom_duration(flunet, subset = NULL,
  remove_warnings = NULL, covariates = NULL)

empirical_distribution_time_to_report(flunet, what = c("symptom_start",
  "symptom_end"), subset = NULL, remove_warnings = NULL,
  covariates = NULL)
}
\arguments{
\item{subset}{character. Logical expression indicating reports to keep: missing values are taken as false. If present, only episode of illness that verify \code{subset} will be processed. This is mainly to set conditions on the warnings to remove unreliable episodes. See \code{remove_warnings} for an alternative way to subset the weekly survey.}

\item{remove_warnings}{character vector. Names of the warnings to exclude.}

\item{covariates}{character vector of covariates to keep.}

\item{what}{character. Specify what date is reported.}

\item{flunet}{a \code{\link{flunet}} object}
}
\value{
A \code{\link{flunet}} object with the empirical distributions stored as list of 3 elements in \code{flunet$log$empirical_distributions}. Each list contains
\itemize{
	\item \code{subset} the logical condition used for subsetting the \code{episode} survey. Made from \code{subset} and \code{remove_warnings} arguments.
	\item \code{covariates} copy of the \code{covariates} argument.
	\item \code{distribution} a \code{data.frame} containing the empirical distribution.
}
}
\description{
These functions compute several empirical distributions:
\itemize{
	\item \code{empirical_distribution_symptom_duration} returns the empirical distribution of the time between symptom start and symptom end dates.
	\item \code{empirical_distribution_time_to_report} returns the empirical distribution of the time between the symptom start or end dates (as specified by the \code{what} argument) and the report date.
}
These distribution are computed from a subset of the \code{episode} survey. If this survey is not present, it will be created from the weekly survey using the function \code{\link{summarize_episode}}.
Subsetting is made via the \code{subset} and/or \code{remove_warnings} arguments. See argument description below.
Finally, covariates to be returned are specified with \code{covariates}. See example below.
}
\examples{
\dontrun{
# symptom duration
bad_warnings <- c("W_S_start_too_far", "W_S_start_before_previous_report", "W_S_start_after_S_end", "W_S_start_wrong","W_S_end_too_far", "W_S_end_before_previous_report", "W_S_end_wrong")
covariates <- "symptom_severity"
flunet <- empirical_distribution_symptom_duration(flunet, remove_warnings=bad_warnings, covariates=covariates)
# time to report symptom start
bad_warnings <- c("W_S_start_too_far", "W_S_start_before_previous_report", "W_S_start_after_S_end", "W_S_start_wrong")
covariates <- c("symptom_severity","first_report_date")
flunet <- empirical_distribution_time_to_report(flunet, what="symptom_start", remove_warnings=bad_warnings, covariates=covariates)
# symptom duration
bad_warnings <- c("W_S_end_too_far", "W_S_end_before_previous_report", "W_S_start_after_S_end", "W_S_end_wrong")
covariates <- c("symptom_severity","last_report_date")
flunet <- empirical_distribution_time_to_report(flunet, what="symptom_end", remove_warnings=bad_warnings, covariates=covariates)
}
}

