% Generated by roxygen2 (4.0.1): do not edit by hand
\name{empirical_distribution_symptom_duration}
\alias{empirical_distribution_symptom_duration}
\title{Compute empirical distribution of symptom duration}
\usage{
empirical_distribution_symptom_duration(flunet, subset = NULL,
  remove_warnings = NULL, summarize_covariates = NULL)
}
\arguments{
\item{subset}{character. Logical expression indicating reports to keep: missing values are taken as false. If present, only episode of illness that verify \code{subset} will be processed. This is mainly to set conditions on the warnings to remove unreliable episodes. See \code{remove_warnings} for an alternative way to subset the weekly survey.}

\item{remove_warnings}{character vector. Names of the warnings to exclude.}

\item{summarize_covariates}{named character vector. Specify how to summarize covariates for each episode of illness. The names correspond to the summarized covariate variable in the outputted data frame.}

\item{flunet}{a \code{\link{flunet}} object}
}
\value{
A \code{\link{flunet}} object with the empirical distribution stored as a \code{data.frame} in \code{flunet$log$empirical_distributions$symptom_duration}.
}
\description{
This function computes the empirical distributions of symptom duration from a subset of the weekly survey.
The subset can be set either with \code{subset} or with a vector of warnings to be removed (\code{remove_warnings}).
Finally, covariates can be summarized for each symptom duration with \code{summarize_covariates}. See example below.
}
\note{
This function takes \code{symptom_start} and \code{symptom_end} from the first and last report respectively. Thus, only episodes with both information are processed.
}
\examples{
\dontrun{
bad_warnings <- c("W_S_start_too_far", "W_S_start_before_previous_report", "W_S_start_after_S_end", "W_S_start_wrong","W_S_end_too_far", "W_S_end_before_previous_report", "W_S_end_wrong")
summarize_severity <- c("symptom_severity"="max(na.omit(symptom_severity))")
empirical_distribution_symptom_duration(flunet, remove_warnings=bad_warnings, summarize_covariates=summarize_severity)
}
}

