% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flunet}
\alias{flunet}
\title{Constructor for flunet object}
\usage{
flunet(intake_survey = NULL, weekly_survey = NULL, contact_survey = NULL,
  context = NULL)
}
\arguments{
\item{intake_survey}{data frame}

\item{weekly_survey}{data frame}

\item{contact_survey}{data frame}

\item{context}{list}
}
\description{
Constructor for flunet object
}

