% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_context_valid}
\alias{is_context_survey_valid}
\alias{is_context_valid}
\alias{is_valid_format}
\alias{is_valid_from_names}
\alias{is_valid_to_names}
\title{Validate context}
\usage{
is_valid_from_names(x = NULL)

is_valid_to_names(x = NULL, survey = NULL)

is_valid_format(x = NULL)

is_context_survey_valid(context = NULL, survey = NULL)

is_context_valid(context)
}
\arguments{
\item{x}{a context variable}

\item{survey}{character, name of the survey}

\item{context}{flunet context}
}
\value{
\code{TRUE} if valid, \code{FALSE} otherwise.
}
\description{
These functions check and validate the fields of the context, the names of the surveys and the names of the survey variables.
}
\seealso{
\code{\link{create_context_flusurvey_201213}} for an example of how to create a flunet context,
\code{\link{variable_context}} for how to define a variable in the context.
}

