% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_variable_name_valid}
\alias{is_variable_name_valid}
\title{Test validity of a variable name}
\usage{
is_variable_name_valid(var_name = NULL, survey = NULL)
}
\arguments{
\item{var_name}{character vector}

\item{survey}{character, the survey where to look at the variable name}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function test whether a variable name is valid and throw a warning otherwise. If more than one \code{var_name} and/or \code{survey} is passed, validity of all names is checked by looking at all the surveys.
}
\examples{
\dontrun{
is_variable_name_valid(c("not_valid","person_id"),survey="intake")
}
}
\seealso{
\code{\link{valid_variable_names}}
}

