% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parsed_names}
\alias{parsed_names}
\title{Transform variable values into names for parsing}
\usage{
parsed_names(x, var_TRUE = NULL, var_FALSE = NULL)
}
\arguments{
\item{x}{a vector of values to be renamed.}

\item{var_TRUE,var_FALSE}{if \code{x} is logical, these are names to replace \code{TRUE} and \code{FALSE} values.}
}
\description{
Transform variable values into names for parsing
}

