% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predicted distribution}
\alias{predicted distribution}
\alias{predicted_distribution_baseline_healthscore}
\alias{predicted_distribution_symptom_duration}
\title{Compute predicted distribution}
\usage{
predicted_distribution_symptom_duration(flunet,
  form = "symptom_duration~symptom_severity+age_group",
  heteroscedasticity = "age_group", range_response = c(0, 20),
  remove_outliers = FALSE, plot = FALSE)

predicted_distribution_baseline_healthscore(flunet,
  form = "baseline_health_score~age_group+smoke+any_UHC",
  heteroscedasticity = "age_group", remove_outliers = FALSE, plot = FALSE)
}
\arguments{
\item{remove_outliers}{logical, if \code{TRUE} then a first regression is run and used to eliminate the outliers before running a second regression. Default to \code{FALSE}.}

\item{flunet}{a \code{\link{flunet}} object}

\item{form}{character, a formula.}

\item{heteroscedasticity}{character, name of the variable used to adjust heteroscedasticity.}

\item{range_response}{numeric vector, range of the response variable (before transformation) to include in the regression. Bounds are not included.}

\item{plot}{logical, if \code{TRUE} several check are plotted. All these plots are also returned.}
}
\value{
A \code{\link{flunet}} object with the predicted distributions stored as list of 6 elements in \code{flunet$log$predicted_distribution}. Each list contains
\itemize{
	\item \code{form}, \code{heteroscedasticity}, \code{range_response}, \code{remove_outliers}: copy of the corresponding arguments.
	\item \code{reg} a \code{list} returned by \code{\link{lme_transform}}.
	\item \code{dist} a \code{data.frame} containing the predicted distribution.
}
}
\description{
These functions compute several predictive distributions:
\itemize{
	\item \code{predicted_distribution_symptom_duration} returns the predicted distribution of the time between symptom start and symptom end dates.
	\item \code{predicted_distribution_baseline_healthscore} returns the predicted distribution of the baseline health-score.
}
These function use linear (potentially mixed effect) models to adjust the response variable to several covariates (specified by the \code{form} argument) and returns the fitted distribution (using predictive confidence intervals) for each combination of the covariate levels.
}
\note{
The response is normalised using either a boxcox or logit transformation.
}
\seealso{
\code{\link{lme_transform}}
}

