% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predicted_distribution_symptom_duration}
\alias{predicted_distribution_symptom_duration}
\title{Compute predicted distribution}
\usage{
predicted_distribution_symptom_duration(flunet,
  form = "symptom_duration~symptom_severity+age_group",
  heteroscedasticity = "age_group", range_response = c(0, 20),
  remove_outliers = FALSE, plot = FALSE)
}
\arguments{
\item{remove_outliers}{logical, if \code{TRUE} then a first lme is run and used to eliminate the outliers before running a second lme. Default to \code{FALSE}.}

\item{flunet}{a \code{\link{flunet}} object}

\item{form}{character, a formula.}

\item{heteroscedasticity}{character, name of the variable used to adjust heteroscedasticity.}

\item{range_response}{numeric vector, range of the response variable (before transformation) to include in the regression.}

\item{plot}{logical, if \code{TRUE} several check are plotted. All these plots are also returned.}
}
\value{
toto
}
\description{
This function use a linear mixed effect (lme) model to adjust the symptom duration to several covariates (specified by the \code{form} argument) and returns the fitted distribution (using predictive confidence intervals) for each combination of the covariate levels.
}
\note{
The symptom duration is normalised using a boxcox transformation.
}
\seealso{
\code{\link{lme_boxcox}}
}

