% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print helper}
\alias{df_print_check}
\alias{df_print_check_raw}
\alias{df_print_define_same_bout}
\alias{print helper}
\title{print helper}
\usage{
df_print_define_same_bout(df, id = NULL)

df_print_check(df, id = NULL)

df_print_check_raw(df, id = NULL)
}
\arguments{
\item{df}{data frame}

\item{id}{character}
}
\description{
print helper
}

