% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resolve_missing_still_ill}
\alias{resolve_missing_still_ill}
\title{Resolve NA in still_ill}
\usage{
resolve_missing_still_ill(df_weekly, my_warning = "W_same_S_start_diff_bout",
  delay_in_reporting = 10, debug = FALSE)
}
\arguments{
\item{df_weekly}{\code{data.frame} containing weekly survey data for a single participant.}

\item{my_warning}{character, warning name}

\item{delay_in_reporting}{maximum number of days to report a date of \code{symptom_start} and \code{symptom_end}. Above this delay, a warning is put on the report.}

\item{debug}{logical, if \code{TRUE}, print checks.}
}
\description{
Resolve missing answer to the question "Are you still ill?"
}

