% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resolve_multiple_profiles}
\alias{resolve_multiple_profiles}
\title{Resolve multiple profiles}
\usage{
resolve_multiple_profiles(flunet)
}
\arguments{
\item{flunet}{a \code{\link{flunet}} object}
}
\description{
Multiple profiles arise when participants update their profile during the season. This function forces only one profile per participant by summarizing logical variables with \code{\link[base]{any}} and non-logical value with \code{\link[dplyr]{last}}. See note below for more details.
}
\note{
\itemize{
	\item For logical variables, \code{\link[base]{any}} is evaluated with \code{na.rm=TRUE}. If all values are \code{NA}, \code{NA} is returned.
	\item For non-logical variables, \code{\link[dplyr]{last}} is evaluated after \code{\link[stats]{na.omit}}. If all values are \code{NA}, \code{NA} is returned.
}
}

