% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sample missing data}
\alias{sample missing data}
\alias{sample_missing_baseline_health_score}
\title{Sample missing data}
\usage{
sample_missing_baseline_health_score(flunet, df_dist)
}
\arguments{
\item{df_dist}{a \code{data.frame} containing the sampling distribution. Must contain at least two columns, one for the value of the sampled variable (with the same name as in \code{flunet} object) and the other for the probability of this value (with names \code{freq}).}

\item{flunet}{a \code{\link{flunet}} object}
}
\value{
A \code{\link{flunet}} object with missing value replaced by sampled ones in the appropriate surveys. A warning column is added to the survey and flagged to \code{TRUE} whenever there is a sampled value.
}
\description{
These functions sample missing data values from a user provided distribution:
\itemize{
	\item \code{sample_missing_baseline_health_score} samples a baseline health-score from the distribution truncated to the highest "ill" health-score.
}
If the distribution is conditioned on some covariates, the sampling is conditioned on the same covariates. Example of such distributions are the predicted distributions (see links below).
}
\seealso{
\code{\link{predicted_distribution_symptom_duration}}, \code{\link{predicted_distribution_baseline_healthscore}}, \code{\link{empirical_distribution_symptom_duration}}, \code{\link{empirical_distribution_time_to_report}}
}

