% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sample_missing_symptom_start_end_dates}
\alias{sample_missing_symptom_start_end_dates}
\title{Sample missing and warning symptom start/end dates}
\usage{
sample_missing_symptom_start_end_dates(flunet, symptom_duration,
  time_to_report = NULL, sample_warning_start = NULL,
  sample_warning_end = NULL)
}
\arguments{
\item{symptom_duration}{data.frame. Predicted distribution of symptom duration, as computed by \code{\link{predicted_distribution_symptom_duration}}}

\item{time_to_report}{data.frame. Empirical distribution of time to report start dates, as computed by \code{\link{empirical_distribution_time_to_report}}}

\item{sample_warning_start}{character vector. Resample start dates with those warnings.}

\item{sample_warning_end}{vector character vector. Resample end dates with those warnings.}

\item{flunet}{a \code{\link{flunet}} object}
}
\description{
This function sample missing or warning symptom start and end dates using a distribution for the duration of symptoms as well as a distribution for the time to report symptom onset.
The distribution of symptom duration can be provided with covariates. If so, the sampling distribution will chosen by matching the covariates to those of the participant/episode.
Sampling proceeds in 3 steps:
\enumerate{
	\item Sample missing or warning start dates for episodes having a valid end date. This is done by sampling from the distribution of symptom duration, truncated according to the time of the reports (for instance symptoms duration must exceed the time from the first report to the symptom end date). Then start date is set as end date minus symptom duration.
	\item Sample missing or warning start dates for episodes having a missing or warning end date. This is done by sampling from the distribution of time to report the start date (if \code{time_to_report} is not null), truncated according to the time of the reports (for instance start date cannot occur before the previous report).
	\item Sample missing or warning end dates for episodes having a valid or sampled start date. This is done by sampling from the distribution of symptom duration, truncated according to the time of the reports (for instance symptoms duration must exceed the time from start date to the penultimate report date of the episode). Then end date is set as start date plus symptom duration.
}
}
\seealso{
\code{\link{predicted_distribution_symptom_duration}}, \code{\link{empirical_distribution_time_to_report}}
}

