% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize_age}
\alias{summarize_age}
\title{Summarize age}
\usage{
summarize_age(flunet, breaks, labels = NULL, include.lowest = TRUE,
  right = FALSE, ordered_result = TRUE, ...)
}
\arguments{
\item{flunet}{a \code{\link{flunet}} object}

\item{breaks}{either a numeric vector of two or more unique cut points or a
    single number (greater than or equal to 2) giving the number of
    intervals into which \code{x} is to be cut.}

\item{labels}{labels for the levels of the resulting category.  By default,
    labels are constructed using \code{"(a,b]"} interval notation.  If
    \code{labels = FALSE}, simple integer codes are returned instead of
    a factor.}

\item{include.lowest}{logical, indicating if an \sQuote{x[i]} equal to
    the lowest (or highest, for \code{right = FALSE}) \sQuote{breaks}
    value should be included.}

\item{right}{logical, indicating if the intervals should be closed on
    the right (and open on the left) or vice versa.}

\item{ordered_result}{logical: should the result be an ordered factor?}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function summarizes the age of every participant with age groups. It's mainly a wrapper aroud \code{\link[base]{cut}}.
}
\examples{
\dontrun{
flunet <- summarize_age(flunet, breaks=c(0,18,45,65,Inf), labels=c("0-17", "18-44", "45-64", "65+"))
}
}

