% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize_baseline_health_score}
\alias{summarize_baseline_health_score}
\title{Compute baseline health score}
\usage{
summarize_baseline_health_score(flunet, fun_summarize = median)
}
\arguments{
\item{fun_summarize}{character. Name of the function used to compute baseline health-score. By default, the median function is used.}

\item{flunet}{a \code{\link{flunet}} object}
}
\value{
A \code{\link{flunet}} object where the \code{intake} survey has an extra column \code{baseline_health_score}.
}
\description{
This function takes all health-score of reports without any symptom and use \code{fun_summarize} to summarize them.
}

