% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize_episode}
\alias{summarize_episode}
\title{Summarize episodes of illness}
\usage{
summarize_episode(flunet)
}
\arguments{
\item{flunet}{a \code{\link{flunet}} object}
}
\value{
A \code{\link{flunet}} object with an additional survey \code{episode}.
}
\description{
This functions summarize the weekly survey (longitudinal data) into episodes of illness. Existing variables of the weekly survey are summarized and several new ones are created (e.g. \code{symptom_duration}).
}
\note{
The weekly survey must not have duplicated report dates (for a given participant). If so, this is solved using the function \code{\link{resolve_multiple_report_date}}.
}

