% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize_severity}
\alias{summarize_severity}
\title{Summarize symptom severity}
\usage{
summarize_severity(flunet, severity = c("ARI_ecdc", "ILI_ecdc", "ILI_fever"))
}
\arguments{
\item{severity}{character vector, symptoms ordered by increasing severity. Each symptoms must correspond to a logical variable in the weekly survey.}

\item{flunet}{a \code{\link{flunet}} object}
}
\value{
a \code{\link{flunet}} object with an additional - ordered - variable \code{symptom_severity} in the weekly survey.
}
\description{
This function summarizes symptom severity by taking the highest symptom as ordered in \code{severity}.
}
\seealso{
\code{\link{summarize_symptoms}}
}

