% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize_smoke}
\alias{summarize_smoke}
\title{Summarize smoking habits}
\usage{
summarize_smoke(flunet, new_var = c("smoke_bool"))
}
\arguments{
\item{new_var}{character vector, new variables created to summarize smoking habits. See note below.}

\item{flunet}{a \code{\link{flunet}} object}
}
\description{
This function summarizes the smoking habits of participants by creating new variables.
}
\note{
The following summary variables are currently available:
\itemize{
	\item \code{"smoke_bool"} is a logical variable that indicates whether participant is a smoker or not.
}
}

