% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize_UHC}
\alias{summarize_UHC}
\title{Summarize underlying health conditions}
\usage{
summarize_UHC(flunet, definitions = c("any_UHC"))
}
\arguments{
\item{definitions}{character vector, one or more definitions to summarize underlying health conditions. See note below.}

\item{flunet}{a \code{\link{flunet}} object}
}
\description{
This function summarizes the underlying health conditions of every participant according to \code{definitions}
}
\note{
\code{any_UHC} stands for any underlying health conditions.
}

