% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test}
\alias{test}
\title{Summarize episode information from longitudinal data.}
\usage{
test(df_data, df_profile, symptom_ordered = c("ARI_ecdc", "ILI_ecdc",
  "ILI_fever"), fun_summary_baseline = fun_summary_baseline,
  approx_W_S_start = c("W_S_start_too_far",
  "W_S_start_before_previous_report", "W_S_start_after_S_end",
  "W_S_start_wrong"), approx_W_S_end = c("W_S_end_too_far",
  "W_S_end_before_previous_report", "W_S_start_after_S_end", "W_S_end_wrong"),
  missing_S_start_date = c("sample_using_S_duration",
  "sample_using_time_to_report", "sample_mix", "use_report_date"),
  missing_S_end_date_not_ill = c("sample_using_S_duration",
  "sample_using_time_to_report", "use_report_date"),
  missing_S_end_date_still_ill = c("sample_using_S_duration",
  "use_report_date"), df_dist_time_to_report_S_start = NULL,
  df_dist_time_to_report_S_end = NULL, df_dist_S_duration = NULL,
  df_dist_baseline_health_score = NULL)
}
\description{
Summarize episode information from longitudinal data.
}

