\name{valid_variable_names}
\alias{valid_variable_names}
\title{Valid names for flunet variables}
\usage{
valid_variable_names()
}
\value{
character vector
}
\description{
This function return a list of two vectors with all the
valid names for flunet variables corresponding to the
intake and weekly questionaires.
}

