% Generated by roxygen2 (4.0.1): do not edit by hand
\name{variable_context}
\alias{variable_context}
\title{context of flunet variable}
\usage{
variable_context(from_name = NULL, to_name = NULL, from_values = NULL,
  to_values = NULL, rename = NULL, revalue = NULL, format = c("factor",
  "ordered", "numeric", "date", "character", "logical"), order_date = "ymd")
}
\arguments{
\item{from_name,to_name}{character, old and new variable names. See \code{rename} for alternative argument passing.}

\item{from_values,to_values}{character vector, old and new variable values. See \code{revalue} for alternative argument passing.}

\item{rename}{named character vector of length 1, with new name as value, and old name as name.}

\item{revalue}{named character vector, with new values as values, and old values as names.}

\item{format}{character, indicate the \R format of the variable.}

\item{order_date}{date character vector, format orders to look for ("ymd" by default). See \code{\link[lubridate]{guess_formats}}.}
}
\value{
list
}
\description{
Create a context for a variable with all necessary information to convert it into a flunet object.
}

