% Generated by roxygen2 (4.0.1): do not edit by hand
\name{char2bool}
\alias{char2bool}
\title{char2bool}
\usage{
char2bool(x, var_names, CR_as_TRUE = FALSE, NA_as_FALSE = FALSE)
}
\arguments{
\item{x}{<what param does>}

\item{var_names}{<what param does>}

\item{CR_as_TRUE}{<what param does>}

\item{NA_as_FALSE}{<what param does>}
}
\description{
char2bool
}

