% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean_weekly_survey}
\alias{clean_weekly_survey}
\title{Clean weekly survey}
\usage{
clean_weekly_survey(flunet, subset = NULL, lag_symptom_start = 2,
  delay_in_reporting = 10, CR_as_TRUE = FALSE, plot_check = FALSE,
  n_cores = 1, debug = FALSE)
}
\arguments{
\item{flunet}{a \code{\link{flunet}} object}

\item{subset}{character, logical expression indicating reports to keep: missing values are taken as false. If present, only episode of illness with at least one report that verify \code{subset} will be processed. This is mainly to save time. E.g. to clean only episodes with at least one ARI report, one can uses \code{subset="ARI_ecdc"}.}

\item{lag_symptom_start}{numeric, maximum number of days between two symptom start dates of different reports. Below this threshold, reports are considered to belong to the same episode of illness and are concatenated.}

\item{delay_in_reporting}{maximum number of days to report a date of \code{symptom_start} and \code{symptom_end}. Above this delay, a warning is put on the report.}

\item{CR_as_TRUE}{logical, if \code{TRUE}, CR (can't remember) is replaced by \code{TRUE} and \code{FALSE} otherwise. This choice is required for the question of whether current illness is the same bout as the one reported the previous time.}

\item{plot_check}{logical, if \code{TRUE}, plot checks.}

\item{n_cores}{number of cores for parallelisation. By default no parallelisation (\code{n.cores=1}). If \code{NULL}, set to half the value returned by \code{\link[parallel]{detectCores}}.}

\item{debug}{logical, if \code{TRUE}, print checks.}
}
\value{
a \code{flunet} object
}
\description{
This function performs several checks, resolves them when possible and otherwise flags informative warnings on the corresponding reports.
}
\note{
The option \code{lag_symptom_start} allows us to group reports belonging to the same bout but having different symptom start dates. This happens when participant change their mind from one report to the next.
}

