% Generated by roxygen2 (4.0.1): do not edit by hand
\name{define_same_bout}
\alias{define_same_bout}
\title{This function identifies reports belonging to the same episode of illness.}
\usage{
define_same_bout(df_weekly, subset = NULL, lag_symptom_start = 2,
  delay_in_reporting = 10, CR_as_TRUE = FALSE,
  my_warning = "W_same_S_start_diff_bout", debug = FALSE, debug_id = NULL)
}
\arguments{
\item{df_weekly}{\code{\link{data.frame}} containing weekly survey data}

\item{my_warning}{character, warning name}

\item{debug_id}{character, person_id of participant to debug}

\item{subset}{character, logical expression indicating reports to keep: missing values are taken as false. If present, only episode of illness with at least one report that verify \code{subset} will be processed. This is mainly to save time. E.g. to clean only episodes with at least one ARI report, one can uses \code{subset="ARI_ecdc"}.}

\item{lag_symptom_start}{numeric, maximum number of days between two symptom start dates of different reports. Below this threshold, reports are considered to belong to the same episode of illness and are concatenated.}

\item{delay_in_reporting}{maximum number of days to report a date of \code{symptom_start} and \code{symptom_end}. Above this delay, a warning is put on the report.}

\item{CR_as_TRUE}{logical, if \code{TRUE}, CR (can't remember) is replaced by \code{TRUE} and \code{FALSE} otherwise. This choice is required for the question of whether current illness is the same bout as the one reported the previous time.}

\item{debug}{logical, if \code{TRUE}, print checks.}
}
\description{
This function identifies reports belonging to the same episode of illness.
}
\note{
This function adds a warning when different bouts are considered to belong to the same episode of illness
}

