% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_suitable_symptom_end}
\alias{find_suitable_symptom_end}
\title{Find suitable symptom end date}
\usage{
find_suitable_symptom_end(df, delay_in_reporting, my_warning, debug = FALSE)
}
\arguments{
\item{df}{\code{data.frame}. Weekly survey of a single participant, containing all the reports of a given episode.}

\item{delay_in_reporting}{maximum number of days to report a date of \code{symptom_start} and \code{symptom_end}. Above this delay, a warning is put on the report.}

\item{my_warning}{character, warning name}

\item{debug}{logical, if \code{TRUE}, print checks.}
}
\description{
This function try to find a suitable symptom end date when several are available
}

