% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_suitable_symptom_start}
\alias{find_suitable_symptom_start}
\title{Find suitable symptom start date}
\usage{
find_suitable_symptom_start(df_weekly, to_match, delay_in_reporting, my_warning,
  debug = FALSE)
}
\arguments{
\item{to_match}{\code{data.frame} with two columns: person_id and n_bout (bout number) whose symptom start dates need to be resolved.}

\item{df_weekly}{\code{\link{data.frame}} containing weekly survey data}

\item{delay_in_reporting}{maximum number of days to report a date of \code{symptom_start} and \code{symptom_end}. Above this delay, a warning is put on the report.}

\item{my_warning}{character, warning name}

\item{debug}{logical, if \code{TRUE}, print checks.}
}
\description{
This function try to find a suitable symptom start date when several are available
}

