% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resolve_multiple_profiles}
\alias{resolve_multiple_profiles}
\title{Resolve multiple profiles}
\usage{
resolve_multiple_profiles(flunet)
}
\arguments{
\item{flunet}{a \code{\link{flunet}} object}
}
\description{
Multiple profiles arise when participants update their profile during the season. This function forces only one profile per participant by summarizing logical variables with \code{\link[myRtoolbox]{any_na}} and non-logical value with \code{\link[myRtoolbox]{last_na}}. See note below for more details.
}
\note{
\itemize{
	\item For logical variables, the logical OR is evaluated over non \code{NA} values. If all values are \code{NA}, then \code{NA} is returned.
	\item For non-logical variables, the last non \code{NA} value is returned. If all values are \code{NA}, then \code{NA} is returned.
}
}

