% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sample_missing_symptom_start_end_dates}
\alias{sample_missing_symptom_start_end_dates}
\title{Sample missing symptom start/end dates}
\usage{
sample_missing_symptom_start_end_dates(flunet, symptom_duration,
  time_to_report = NULL, sample_warning_start = NULL,
  sample_warning_end = NULL)
}
\arguments{
\item{symptom_duration}{distribution}

\item{time_to_report}{distributions}

\item{sample_warning_start}{vector}

\item{sample_warning_end}{vector}

\item{flunet}{a \code{\link{flunet}} object}
}
\description{
This function sample missing symptom start and end dates using a distribution of the duration of symptoms
}
\note{
TODO
}

